Split omemo plugin up into multiple files
